+++ packages/cli/src/commands/service/dev.ts
@@ -1,87 +1,111 @@
-import { IService } from '@mesg/api/lib/service'
-import * as base58 from '@mesg/api/lib/util/base58'
+import { Command, flags } from '@oclif/command'
+import { join } from 'path'
+import Listr from 'listr'
+import LCD from '@mesg/api/lib/lcd'
+import API from '@mesg/api'
+import chalk from 'chalk'
+import { decode } from '@mesg/api/lib/util/encoder'
+import { parseLog } from '../../utils/docker'
+import * as Environment from '../../tasks/environment'
+import * as Service from '../../tasks/service'
+import * as Runner from '../../tasks/runner'
+import * as Instance from '../../tasks/instance'
+import version from '../../version'
 
-import Command from '../../root-command'
-import { IsAlreadyExistsError } from '../../utils/error'
+const ipfsClient = require('ipfs-http-client')
 
-import ServiceCompile from './compile'
-import ServiceCreate from './create'
-import ServiceLog from './logs'
-import ServiceStart from './start'
-import ServiceStop from './stop'
+type Context = Environment.IStop | Service.ICompile | Service.ICreate | Runner.ICreate | Runner.ILogs | Instance.IEventLogs | Runner.IResultLogs
+type ContextEnd = Runner.ILogsStop | Instance.IEventLogsStop | Runner.IResultLogsStop | Environment.IStop
 
-export default class ServiceDev extends Command {
-  static description = 'Run a service in development mode'
+export default class Dev extends Command {
+  static description = 'Run a service in a local development environment'
 
   static flags = {
-    ...Command.flags,
-    ...ServiceCreate.flags,
-    ...ServiceStart.flags,
+    image: flags.string({ name: 'Engine image', default: 'mesg/engine' }),
+    tag: flags.string({ name: 'Engine version', default: version.engine }),
+    pull: flags.boolean({ name: 'Force to pull the docker image', default: false }),
+    configDir: flags.string({ name: 'Configuration directory', default: join(process.cwd(), '.mesg') }),
+    configFile: flags.string({ name: 'Config filename', default: 'config.yml' }),
+    env: flags.string({
+      description: 'Environment variables to inject to the service',
+      multiple: true,
+      helpValue: 'FOO=BAR'
+    })
   }
 
   static args = [{
-    name: 'SERVICE',
+    name: 'PATH',
     description: 'Path or url of a service',
     default: './'
   }]
 
-  runnerCreated = false
-
   async run() {
-    const { args, flags } = this.parse(ServiceDev)
+    const { args, flags } = this.parse(Dev)
 
-    this.spinner.start('Starting service')
-    this.spinner.status = 'compiling'
-    const definition = await ServiceCompile.run([args.SERVICE, '--silent', ...this.flagsAsArgs(flags)])
-    this.spinner.status = 'creating'
-    const serviceHash = await this.createService(definition)
-    this.spinner.status = 'starting'
-    const runnerHash = await this.startService(serviceHash, flags.env)
-    this.spinner.status = 'fetching logs'
-    const stream = await ServiceLog.run([runnerHash, ...this.flagsAsArgs(flags)])
-    this.spinner.stop(runnerHash)
+    const tasks = new Listr<Context>([
+      Environment.start,
+      Service.compile,
+      Service.create,
+      Runner.create,
+      {
+        title: 'Start service\'s logs',
+        task: () => new Listr<Runner.ILogs | Instance.IEventLogs | Runner.IResultLogs>([
+          Runner.logs,
+          Instance.eventLogs,
+          Runner.resultLogs,
+        ])
+      }
+    ])
+    const result = await tasks.run({
+      configDir: flags.configDir,
+      configFile: flags.configFile,
+      env: flags.env,
+      image: flags.image,
+      pull: flags.pull,
+      tag: flags.tag,
+      ipfsClient: ipfsClient('ipfs.app.mesg.com', '5001', { protocol: 'http' }),
+      path: args.PATH,
+      grpc: new API('localhost:50052'),
+      lcd: new LCD('http://localhost:1317'),
+    })
+
+    const runnerLogs = (result as Runner.ILogs).runnerLogs || []
+    const eventLogs = (result as Instance.IEventLogs).eventLogs
+    const resultLogs = (result as Runner.IResultLogs).resultLogs
+
+    for (const log of runnerLogs) {
+      log
+        .on('data', buffer => parseLog(buffer).forEach(x => this.log(chalk.gray(x))))
+        .on('error', error => { this.warn('Docker log stream error: ' + error.message) })
+    }
+
+    eventLogs
+      .on('data', event => this.log(`EVENT[${event.key}]: ` + chalk.gray(JSON.stringify(decode(event.data)))))
+      .on('error', error => { this.warn('Event stream error: ' + error.message) })
+
+    resultLogs
+      .on('data', execution => execution.error
+        ? this.log(`RESULT[${execution.taskKey}]: ` + chalk.red('ERROR:', execution.error))
+        : this.log(`RESULT[${execution.taskKey}]: ` + chalk.gray(JSON.stringify(decode(execution.outputs)))))
+      .on('error', error => { this.warn('Result stream error: ' + error.message) })
 
     process.once('SIGINT', async () => {
-      stream.destroy()
-      if (this.runnerCreated) await ServiceStop.run([runnerHash, ...this.flagsAsArgs(flags)])
-    })
-  }
-
-  async createService(definition: IService): Promise<string> {
-    const hash = await this.lcd.service.hash({
-      configuration: definition.configuration,
-      dependencies: definition.dependencies,
-      description: definition.description,
-      events: definition.events,
-      name: definition.name,
-      repository: definition.repository,
-      sid: definition.sid,
-      source: definition.source,
-      tasks: definition.tasks
-    })
-    if (!hash) throw new Error('invalid hash')
-    const exists = await this.lcd.service.exists(hash)
-    if (!exists) {
-      const service = await this.api.service.create(definition)
-      if (!service.hash) throw new Error('invalid hash')
-      if (base58.encode(service.hash) !== hash) throw new Error('invalid hash')
-    }
-    return hash
-  }
-
-  async startService(serviceHash: string, env: string[]): Promise<string> {
-    try {
-      const { hash } = await this.api.runner.create({
-        serviceHash: base58.decode(serviceHash),
-        env
+      await new Listr<ContextEnd>([
+        {
+          title: 'Stopping logs',
+          task: () => new Listr<Runner.ILogsStop | Instance.IEventLogsStop | Runner.IResultLogsStop>([
+            Runner.logsStop,
+            Instance.eventLogsStop,
+            Runner.resultLogsStop
+          ])
+        },
+        Environment.stop,
+      ]).run({
+        configDir: (result as Environment.ICreateConfig).configDir,
+        eventLogs: eventLogs,
+        resultLogs: resultLogs,
+        runnerLogs: runnerLogs,
       })
-      if (!hash) throw new Error('invalid hash')
-      this.runnerCreated = true
-      return base58.encode(hash)
-    } catch (e) {
-      if (!IsAlreadyExistsError.match(e)) throw e
-      this.warn('service already started')
-      return base58.encode(new IsAlreadyExistsError(e).hash)
-    }
+    })
   }
 }