+++ packages/cli/src/commands/service/doc.ts
@@ -1,51 +1,46 @@
-import { flags } from '@oclif/command'
-import { readdirSync, readFileSync, writeFileSync } from 'fs'
-import { compile, registerHelper } from 'handlebars'
-import { safeLoad } from 'js-yaml'
+import { flags, Command } from '@oclif/command'
+import { readdirSync, writeFileSync } from 'fs'
 import { join } from 'path'
+import Listr from 'listr'
+import * as Service from '../../tasks/service'
 
-import Command from '../../root-command'
+const ipfsClient = require('ipfs-http-client')
 
-export default class ServiceDoc extends Command {
-  static description = 'Generate documentation for service and print it on stdout'
+type Context = Service.ICompile | Service.IGenDock
+
+export default class Doc extends Command {
+  static description = 'Generate documentation for service and print it to stdout'
 
   static flags = {
-    ...Command.flags,
     save: flags.boolean({ char: 's', description: 'Save to default readme file' }),
   }
 
   static args = [{
-    name: 'SERVICE',
+    name: 'SERVICE_PATH',
     description: 'Path of a service',
     default: './'
   }]
 
   async run(): Promise<string> {
-    const { args, flags } = this.parse(ServiceDoc)
-    const definition = safeLoad(readFileSync(join(args.SERVICE, 'mesg.yml')).toString())
-    const markdown = this.generateTemplate(definition)
+    const { args, flags } = this.parse(Doc)
+    const tasks = new Listr<Context>([
+      Service.compile,
+      Service.genDoc
+    ])
+    const result = await tasks.run({
+      ipfsClient: ipfsClient('ipfs.app.mesg.com', '5001', { protocol: 'http' }),
+      path: args.SERVICE_PATH,
+    })
+    const markdown = (result as Service.IGenDock).markdownDoc
     if (flags.save) {
-      this.saveReadme(args.SERVICE, markdown)
+      const defaultReadmeFileName = readdirSync(args.SERVICE_PATH).find(file => {
+        return /^readme(?:.(?:md|txt)+)?$/i.test(file)
+      })
+      const readmeFileName = defaultReadmeFileName || 'README.md'
+      writeFileSync(join(args.SERVICE_PATH, readmeFileName), markdown)
     } else {
       this.log(markdown)
     }
     return markdown
-  }
-
-  generateTemplate(data: any) {
-    registerHelper('or', (a: any, b: any) => a ? a : b)
-    registerHelper('toJSON', function (obj) {
-      return JSON.stringify(obj);
-    });
-    const template = readFileSync(join(__dirname, '..', '..', '..', 'assets', 'doc.md')).toString()
-    return compile(template)(data)
-  }
-
-  saveReadme(servicePath: string, markdown: string) {
-    const defaultReadmeFileName = readdirSync(servicePath).find(file => {
-      return /^readme(?:.(?:md|txt)+)?$/i.test(file)
-    })
-    const readmeFileName = defaultReadmeFileName || 'README.md'
-    writeFileSync(join(servicePath, readmeFileName), markdown)
   }
 }