+++ packages/cli/src/commands/service/execute.ts
@@ -1,17 +1,26 @@
-import {flags} from '@oclif/command'
-import {readFileSync} from 'fs'
-import {ExecutionCreateOutputs} from '@mesg/api/lib/execution'
-import * as base58 from '@mesg/api/lib/util/base58'
+import { flags, Command } from '@oclif/command'
+import { readFileSync } from 'fs'
+import { ExecutionCreateOutputs } from '@mesg/api/lib/execution'
+import { cli } from 'cli-ux'
+import LCD from '@mesg/api/lib/lcd'
+import Application from '@mesg/application'
+import API from '@mesg/api'
+import Listr from 'listr'
+import * as Execution from '../../tasks/execution'
+import * as Runner from '../../tasks/runner'
+import * as Service from '../../tasks/service'
+import * as Instance from '../../tasks/instance'
+import { IService } from '@mesg/api/lib/service-lcd'
+import ServiceType from "@mesg/api/lib/typedef/service";
 
-import Command from '../../root-command'
-import {runnerResolver} from '../../utils/resolver'
+export type IValidateTask = { service: IService, taskKey: string, task?: ServiceType.mesg.types.Service.ITask }
+export type Context = Runner.IGet | Instance.IGet | Service.IGet | IValidateTask | Execution.IConvertInput | Execution.IExecute
 
-export default class ServiceExecute extends Command {
+export default class Execute extends Command {
   static description = 'Execute a task on a running service'
 
   static flags = {
-    ...Command.flags,
-    json: flags.string({char: 'j', description: 'Path to a JSON file containing the task inputs'}),
+    json: flags.string({ char: 'j', description: 'Path to a JSON file containing the task inputs' }),
     data: flags.string({
       char: 'd',
       description: 'Task inputs',
@@ -34,70 +43,45 @@ export default class ServiceExecute extends Command {
   }]
 
   async run(): ExecutionCreateOutputs {
-    const {args, flags} = this.parse(ServiceExecute)
+    const { args, flags } = this.parse(Execute)
+    const app = new Application(new API(`localhost:50052`))
 
-    const runnerHash = await runnerResolver(this.lcd, args.RUNNER_HASH)
-
-    const runner = await this.lcd.runner.get(runnerHash)
-    if (!runner.instanceHash) { throw new Error('invalid runner hash') }
-
-    const instance = await this.lcd.instance.get(runner.instanceHash)
-    if (!instance.serviceHash) { throw new Error('invalid instance') }
-
-    const service = await this.lcd.service.get(instance.serviceHash)
-    if (!service.hash) { throw new Error('invalid service') }
-
-    const task = service.tasks.find(x => x.key === args.TASK)
-    if (!task) {
-      throw new Error(`The task ${args.TASK} does not exist in service '${service.hash}'`)
-    }
-
-    const result = await this.execute({
-      inputs: this.convertValue(task.inputs, this.dataFromFlags(flags)),
-      executorHash: base58.decode(runnerHash),
-      tags: ['CLI'],
+    const tasks = new Listr<Context>([
+      {
+        title: 'Validate task',
+        task: () => new Listr<Runner.IGet | Instance.IGet | Service.IGet | IValidateTask>([
+          Runner.get,
+          Instance.get,
+          Service.get,
+          {
+            title: 'Validate task',
+            task: (ctx: IValidateTask) => {
+              ctx.task = ctx.service.tasks.find(x => x.key === ctx.taskKey)
+              if (!ctx.task) throw new Error(`The task ${ctx.taskKey} does not exist in service '${ctx.service.hash}'`)
+            }
+          }
+        ])
+      },
+      Execution.convertInput,
+      Execution.execute,
+    ])
+    const res = await tasks.run({
+      lcd: new LCD(`http://localhost:1317`),
+      runnerHash: args.RUNNER_HASH,
       taskKey: args.TASK,
-      eventHash: base58.decode(flags.eventHash || '6aUPZhmnFKiSsHXRaddbnqsKKi9KogbQNiKUcpivaohb'), // TODO: to improve
+      app: app,
+      data: flags.json
+        ? JSON.parse(readFileSync(flags.json).toString())
+        : (flags.data || []).reduce((acc, item) => ({
+          ...acc,
+          [item.split('=')[0]]: item.split('=')[1]
+        }), {}),
+      eventHash: flags.eventHash || '6aUPZhmnFKiSsHXRaddbnqsKKi9KogbQNiKUcpivaohb',
+      tags: ['CLI'],
     })
-    this.styledJSON(result)
+
+    const result = app.decodeData((res as Execution.IExecute).result.outputs)
+    cli.styledJSON(result)
     return result
-  }
-
-  dataFromFlags(flags: { data: string[], json: string | undefined }): any {
-    if (flags.json) {
-      return JSON.parse(readFileSync(flags.json).toString())
-    }
-    return (flags.data || []).reduce((acc, item) => {
-      const [key, value] = item.split('=')
-      return {
-        ...acc,
-        [key]: value
-      }
-    }, {})
-  }
-
-  convertValue(inputs: any, data: any): any {
-    if (!inputs) return {}
-    return inputs
-      .filter((x: any) => data[x.key] !== undefined && data[x.key] !== null)
-      .reduce((prev: any, value: any) => ({
-        ...prev,
-        [value.key]: this.convert(value.type, data[value.key])
-      }), {})
-  }
-
-  convert(type: 'Object' | 'String' | 'Boolean' | 'Number' | 'Any', value: string | any): any {
-    try {
-      return {
-        Object: (x: string | any) => typeof x === 'string' ? JSON.parse(x) : x,
-        String: (x: string) => x,
-        Boolean: (x: string) => ['true', 't', 'TRUE', 'T', '1'].includes(x),
-        Number: (x: string) => parseFloat(x),
-        Any: (x: string) => x,
-      }[type](value)
-    } catch {
-      this.warn(`Cannot parse ${value} in ${type}`)
-      return value
-    }
   }
 }